ARM: tegra: power: Apply down delay to balancing CPUs
On Tegra3 secondary G-CPU may be turned off by auto-hotplug governor
in two cases: when overall CPU load is low enough to justify transition
to LP CPU, or when CPU cores usage by the scheduler is unbalanced
(skewed). In the former case down delay (currently 2sec) was inserted
before the core is turned Off. In the latter case the up delay (100ms)
was used, i.e., the same delay applied to balancing cores regardless
of the On/Off direction.

This commit would apply down delay when turning core Off in both cases
above, and keep using up delay only for turning core On.

Change-Id: I3ac01c4df833eaf020966d561ec4b9802e25b529
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/99737
Tested-by: Jay Cheng <jacheng@nvidia.com>
Reviewed-by: Prajakta Gudadhe <pgudadhe@nvidia.com>