ARM: tegra: cardhu: wakeup system from GPIO_PV0 without key detection
To meet the LP0 exit power on sequence, it is require to wake system
for tegra gpio in place of PMIC for E1291-A04.
Also it is observed that if GPIO key is used to wakeup then there is
possibility of loosing the key event and hence adding the gpio
GPIO_PV0 as the key with code of RESERVED so that it can only
wakeup system but will not able to send the key event through
gpio keys.

bug 981320

Change-Id: I8610adca4b5ed8ae79f8fcca9a1d4b5548158c60
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/101784
Tested-by: Sang-Hun Lee <sanlee@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User