media: nvavp: make the clock requests be reference counted
Previously we only managed the clocks around pushbuffer submit by using
the non-emptiness of the pushbuffer as an implicit refcnt on the clock.
Add an actual refcnt on the clocks to allow other entities to keep
references on the clocks. Still use the queue being non-empty to request
clock, now represented by the pending flag.

This is a prerequisite for a follow-on patch that will allow avp client
to keep extra refs on the clocks to workaround an issue where the bsev
clocks need to be on to erase the DRM keys AFTER decode has been
finished. Currently, this would cause a lockup since the clocks would
be turned off at the end of decode.

Change-Id: I8314657d0073614ab6a2f6708ed1785e9aff2ff2
Reviewed-on: http://git-master/r/102023
Reviewed-by: Karan Jhavar <kjhavar@nvidia.com>
Reviewed-by: Prajakta Gudadhe <pgudadhe@nvidia.com>
Tested-by: Prajakta Gudadhe <pgudadhe@nvidia.com>

Conflicts:

	drivers/media/video/tegra/nvavp/nvavp_dev.c